+++ lib/honeybadger/agent.rb
@@ -309,7 +309,6 @@ module Honeybadger
     # @!method exception_filter
     # @yieldreturn [Boolean] true (to ignore) or false (to send).
     def exception_filter(&block)
-      warn 'DEPRECATED: Honeybadger.exception_filter is deprecated. Please use Honeybadger.before_notify instead.'
       config.exception_filter(&block)
     end